requirements.txt
google-generativeai
python-dotenv
SpeechRecognition 
gTTS 
playsound
pvporcupine
pyaudio